Refine Docker Hub notifications
The Docker Hub notification has an images field that can be either null or an
array of image ids that were pushed. In the case of an automated build with N
tags, the webhook will receive N-1 notifications with images=null and a final
notification with an N element array.

The new notification logic is:

* images=null || images.length === 0: No notification
* images.length === 1: Notification including short id of pushed image
* images.length > 1: Notification including number of images pushed